Add get_prefix() function
This function will try as hard as it can to always register and then
return a prefix to the user. It will use find_free_prefix() in a loop
to find a free prefix, insert it into the database and if it fails, it
will retry with a new free prefix, thus avoiding race conditions.