Add indent to table
This is used in the presentation layer to give a nicer overview of
prefixes. Since the database is lots-of-read and not so many writes
together with indent-calc being somewhat heavy, we choose to store the
indent level in the db and not calculate it on the fly when reading the
data.