Update SQL schema
We now have both a prefix and a display_prefix column. The prefix column
is used to store the actual prefix, make sure there are no collisions
and calculate indentation while display_prefix is what the user will
typically want to see.

For a host in a /27 network, the prefix column would contain the address
with a /32 mask while display_prefix would contain it with it's natural
/27 mask. If the prefix column would have a /27 mask we would not be
able to recognize that this host is a child of the /27 network nor would
we be able to register two /32 hosts under a /31 if we used the real
netmask since the first host and the network address would be the same.
With this approach the network is stored as a /31 while the first host
is a /32, we thus achieve uniqueness. display_prefix is updated using
triggers together with the indent level. This information could be
calculcated on the fly, but due to the relative few write operations
compared to reads we choose to store this extra data for faster reads.