More tests and error handling
Added some tests for input testing (something like fuzzy testing) to
some functions, just schema related for now.

NAP has got a new function _check_attr to check that a given dict
contains the required attributes, no more than the allowed_attrs or
it will throw an error.

The spec needs to be checked in a better manner as well.