Fuglify database
Remove a pretty useless table used just to store default prefix-length
for a pool. Since it can differ between IPv4 and IPv6 we had a separate
table to store it.

The rationale behind having the table is that it's more elegant from a
strict database perspective but boy, is it a pain in the ass to work
with (yes it is!). We now scrap the table and put it as two extra
columns instead.