Fix test cases for edit_pool
We fix the edit_pool test, which purpose really is to try to edit a pool
to use id instead of name since that is not a legal combination (must be
specified by schema and name).

In addition, we create a new test which tries to edit a pool just by
name, which isn't a legal combination and should raise an exception.