Add some more tests for find_free_prefix
One merely checks that ffp can return "a lot" of prefixes (256 in this
case)... the other that we can't specify a too high "wanted number of
returned prefixes". Max should be 1000 and the stored procedure will
raise an exception whenever over that, but it seems we fail to pick up
that exception and so this test fails.