Another hack :(
This just proved we need a better way to adapt a python list of IP
prefixes to a proper postgresql array of inets.

I can't even explain (although I'm tired and on a flight) why this is
happening. If we input a list containing two prefixes to
find_free_prefix in its from-prefix argument.. it works! If we just give
it one argument, it doesn't work. PostgreSQL says it doesn't know what
the type the object is and it can't cast from unknown to inet, so now we
haxx it and first cast to text then to inet.. sigh, so ugly..