Rework find_free_prefix args
Instead of taking a spec (which isn't really a schema_/pool_/prefix_
spec) and a couple of other arguments, such as count and
wanted_prefix_length, we now take a dict called args containing all of
that. This provides greater flexibility for passing things such as
address-family and prefix-length (it's optional when using
'from-pool').. and so forth... we keep two arguments at the end for
compatibility reasons, as soon as the rest of the code is updated, we'll
remove it!