find_free_prefix() now supports from-pool!
It's bloody working, we can now ask to get a prefix from a pool instead
of just from a list of prefixes. Put in some unittests for it as well...
seems like most things are working correctly.

Also introduced a new exception, NapNonExistentError which is raised
when a non-existent object, such as a pool, is referenced and does not
exist.