Improvements to SQL testing
This test is meant to provide more direct testing of the SQL database
related things such as the indentation calculation and different
constraints. It uses SQL statements directly although they are executed
through nap (using _execute()) so that the nap class can translate
exceptions for us... that might change in a later release.