Added display attr. to result from search_prefix.
The prefixes returned from search_prefix (and therefore also from
smart_search_prefix) now contain a 'display' attribute. This is used
when we want to return a complete data set for simplicity, but the user
has chosen not to display the entire result set.

Fixed a small but in the xmlrpc_remove_prefix function.