Improve smart search
smart search would misinterpret certain things, such as 1.3 (which is a
valid representation of the IPv4 address 1.0.0.3) before, this is no
longer the case as it makes search it has all four octets of an IPv4
address before it treats it as a prefix.

Further, searching for a prefix such as 1.3.3.0/24 would not return
1.3.3.1 if 1.3.3.0/24 did not exist (and you had immediate children
display option set). Since we want to be able to find all prefixes
within another, we now automatically perform a "contained_within_equals"
search when we see a prefix. This could potentially lead to a lot of
results, but we hope to handle that by pagination and other means.