Add "Add prefix" button & fix CSS
There's now a link on the "prefixes" page (really the prefix_list.html)
that links to the page for adding a new prefix. Essential for
navigation! ;)

CSS has been fixed in a lot of ways, some tabs inserted by last commit
was changed to whitespaces. Some gradient backgrounds now work in Chrome
as well as in Firefox as well as some border-radiuses and shadows.