Fix speedtest output
This fixes the output of the network count of speedtest. Output is of
the format:
[network] [time]
You get one line per /24 inserted and so each network row represent the
average time it took to insert one prefix in that /24. The value for
network was previously taken from the third octet of the inserted IP
address which would work fine for one /24 but when you insert multiple
/24s the network value would wrap and start over at 0.

Now network is a free flowing counter that starts at 0 and just
increases.