Improved display of error messages
Error messages are now displayed in a nice popup-div instead of the
temporary JavaScript alert()-solution. The error notice is was also
unified with the verification dialog which existed previously.

A separate JavaScript file was also created, now containing the error
handling and dialog box functions. In the mean time more and more
functions will be moved there.