Moved JavaSript functions to external file
Moved most JavaScript functions regarding the prefix search and prefix
list to the nap.js-file. Also restructured a bit; instead of having two
very similar functions to add prefixes to the list, there is now only
one.

Also made a really ugly hack to be able to use the same function for
adding prefixes to the list on both the add prefix and prefix list
pages. As the links on the prefixes needs to be different in the two
cases, a global variable was introduced to keep state of what type of
link to place on the current page.
*sigh*
Hopefully a better solution well emerge in the future.