Vertically center text on button
This is a bit tricky, not only because how it renders will likely differ
from browser to browser, but because some of our buttons are based on
submit (to submit a form) while others are mere a href links. The submit
buttons have som form of padding so with the same line-height property,
it looks centered on a div button and non-centered on a submit button.
For this we now have a separate case for submit buttons which sets a
different line-height.

Guess this somewhat affescts issue #1.