Improve search interpretation information
Backend now includes an "expanded" value, in case the value sent in the
query string was expanded. The only instance when this happens today is
when shortening an IPv4 prefix, such as 10/8 which is automatically
expanded to 10.0.0.0/8 in the backend. To let the user know it was
expanded, it is now included in the search interpretation data structure
and the web interface checks for its existance and displays if it finds
it.

Part of issue #20