Add support for using ip4r index in prefix search
This hacks in support for using ip4r index on a prefix search when the
contains / contained within operators are used which speeds up prefix
searches tremendously. It is currently only being performed on the
"inner" query, but that is also the one that in a "classic" case takes
the longest.

Test suit passes.

Part of issue #7