An alternative calc_indent
This is an alternative approach to calc_indent.

It passes the test suite and to me, it looks a bit cleaner and easier to
understand. Just cleaning up the old version and indenting it properly
might make that easy enough though.

There is no performance improvement with this version. Empirical tests
show it is on par and theoretically I don't see that it should be either
faster nor slower. A larger dataset and a different decision by the
query planner could of course lead to a larger delta in performance
between the two versions.

Related to #25, for which I originally wrote it in hope of better
performance just to realize it was not the bottleneck in the first
place.