Avoid unnecessary redraw in prefix search
The query input box detects changes and emits a new search. That would
include just moving the cursor in the box and since the actual search
string would be unchanged, a query equal to the currently rendered (or
outstanding, in case we haven't received the answer yet) would be
emitted.

With this change we save any sent search in a global variable and on a
change from the query input box we check that it's different from the
global current_query variable. If not, we do nothing.

Fixes #19.