Added timing info and search state tracking
Now timing info is written to a stats-object when a prefix search is
performed (see issue #27). The information is atm just presented in the
JavaScript console, but will be presented to the user in a better way
soon.

To determine if there is a problem with overlapping searches, that is
search result arriving when a previous result is beeing drawn, code was
added to track the state of prefix searches. If the prefix list
functions end up in a state where it is not supposed to be, a message is
printed to the JavaScript console.