Improve text-import
Don't always assume /32s are loopbacks - they can be hosts too. If they
are hosts, they should be covered by a larger prefix of type assignment.
We first try to insert the /32 as an assignment (assuming it's a
loopback), if it indeed is a host, we'll get an exception and we'll
instead try to insert it as a host.

Further, we now insert "other" prefixes, ie stuff not in the range /30 -
/32.

Part of issue #5.