Pool expansion functionally complete
Now pools can be expanded with new prefixes created directly on the pool
edit page in a fashion similar to how ordinary new prefixes are created.
Left to get #2 finished is some cosmetic fixes, such as hiding prefix
attribues which not are needed.

Also fixed a bug in the list_pool API function; the prefix list of an
empty pool (containing no prefixes) will now be an empty list instead of
None.