Add javascript log function
We previously relied directly on console.log, which is only available in
Chrome or if you run Firefox with Firebug. Since that is not always the
case, we've wrapped the whole thing in a new log() function which will
try to send a console.log and thereafter open a Opera error message. If
nothing works it will silently give up and discard the log line.