Include query search_options in result
The search_options submitted together with the query is now returned
together with the result. This is primarily to get back the query_id,
which is a new attribute optionally set by the client to keep track of
multiple outstanding queries.

This will likely break stuff since the return values of search_prefix
has changed.

Part of issue #38