Revamp/polish last change
smart_search_prefix always returned a dict with the result and that
would easily be extendable to include more data in other keys in the
dict. search_prefix on the other hand returned a list with the result,
since we want to be able to receive query_id and potentially other data
revolving around the actual search result, its return format was changed
into a dict, just like with smart_search_prefix.

This also means smart_search_prefix can take the return result straight
from search_prefix (and just add in the interpretation).

Napmodel is obviously changed to accept the new format emitted by the
above change and so is the xhr class and then the javascript for the web
UI.

This fixes #38, although there are likely some things that are curently
broken due to our changes.