Use iprange index for great success
Now use iprange index on the ip_net_plan self-join. This speeds up the
query tremendously. We have previously used ip4r and a hack on selected
queries to speed up IPv4 parts while still leaving IPv6 functional. On
this particular query, we felt that would have resulted in a too ugly
query so we've refrained from it. With iprange it is alllll gooooood!

Part of #6

Naturally, this adds iprange to the list of our dependencies :)