Polish & shine for pagination
Remove a debug log line and lower the threshold for getting the next
page, ie get it earlier to make sure the user "never" hits the bottom.
That is subject to change with screen size. The current setting of 2
works great on a 1024x768 with a fair bit of margin. It will hopefully
look good on 1920x1200 screens too.

Part of #4.