Add 'Display more...' link for prefix results
On the prefix search page we have automatic pagination when you approach
the bottom of the results. This is not suited for some other pages, such
as the add prefix page or edit pool page and for those we instead add a
small link so one gets to paginate manually.

It needs to be hidden once we hit the end of the results.

Part of #4.