Change inactive menu link text color
One of our users were confused just on which page he was currently at. I
showed how Google does the exact same thing we do, except that Google
actually have bold on the active link and normal weight on the rest.
That looks rather silly on our page as the text width shifts so much and
instead I decided to use a light gray to simulate the same effect.