Changes to database constraints
Remove NOT NULL constraints for alarm_priority and monitor since we
should be able to leave them in an unset state. Also remove the defaults
for ip_net_plan type, pool default_type and monitor / alarm_priority.
All that can be handled in python middleware instead. Let's not enforce
default values in the database, just constraints.