Update db & napbase test suite & bug fixes
Mostly due to the removal of certain default values in the database so
we now have to specify more things manually.

This revealed a bug that was introduced with the windowing function -
sorting was lost. We now have an outer ORDER BY too, might be able to
remove the middle one for a slight performance increase.