Prefix list widened and built with tables
The prefix list is widened to utilize more screen area. Currently, the
max-width is set to 1200 pixels. After many failed tries to make the
description text flow nicely, the divs which build up the list was
altered to behave like table elements. To maintain the nice collapse
functionality we have, unfortunately each row in the prefix list is made
as a separate table. Not very pretty, but gets the job done.

It was also realized that one class never was applied to the columns in
the prefix search options, which is why it might look a bit funny. See
separate ticket.

Anyway, completed a major part of #32.