Workaround shlex's lack of support for unicode
shlex, the module we use to split the query string into parts, does not
support Unicode and so this is a workaround for that. This is not a 100%
solution since the UTF8 to 8-bit string conversion likely will fail on
certain text strings although it certainly works better than it has thus
far.

I could likely write a 100% solution, out of a character compatibility
perspective, but the chance if introducing other bugs in the actual
splitting code would increase. The current solution is a compromise
between 100% split correctness and < 100% character compatibility.

Part of #49