Add search help
This adds a nice little popup that explains how the search function is
used. It's displayed using a slightly more generalized dialog system,
although that is not complete yet.

The dialogs now have a slightly different look with a drop shadow and
rounded border, though it looks funky on all dialogs except for the help
one since there's a non-rounded div at the top.