Redesign search box somewhat
This is a revamp of the search box area to make it work better on small
as well as large screens (using min & max-width together with sane
percentual width). It is detached from the search interpretation box so
that the search query input field does not move depending on inputted
search query. Looks nice! :)

Fixes #58