Revamp hidden prefix, crouching code
This modifies the code for hiding prefixes. There was a bug for when
getting the "next" set of prefixes. We now store the container globally
so we can keep state across different result sets.

We now count how many prefixes are in a container and if less than X
(depending on case) we don't hide them.

Fixes #70