New mechanism for rendering duplicate errors
We catch unique index failures in our db execute function and raise our
own exception for it. The actual index name doesn't mean much to the end
user, so we here make an attempt to replace it with the comment stored
in the database on the violated unique index. Looks better!

Fixes #89.