Added pool and schema search functions to NapModel
Added the search and smart_search functions for schemas and pools to the
NapModel. At the same time a major change was done to the return values
of all search functions; previously the search result was placed in
elements named 'prefix_list' or 'result_list'. These were all renamed to
'result'.

Part of #74