Test of DataTables
A test of the DataTables jQuery plugin and its server-side processing
feature. As this feature is quite "narrowly" implemented and for example
can not add a list of objects to a table, only a list of lists, we will
probably not use this plugin. I commit the code anyway, for reference if
we chose to pick it up again.

Related to #74