Toolbox does not mess up prefix row
Bringing up the toolbox menu no longer messes up the prefix row. Fixed
by appending the popup div to the body instead of placing it in the row.
Also modified the prefix list not to create all menus, hidden, when the
page first is rendered and instead creates the divs when needed.

Fixes #98