Blargh, fugly fixes for render logic
This is not a very pretty solution, we should really try to loop
through and only handle each case once, but it's too late to find a
pretty solution right now.

I think this leaves only one more thing to fix #103. When a prefix is
expanded but no childs are found that need to be rendered, ie if all
childs are already rendered, then the expand-button should change from +
to -.. that does not yet work.

This probably needs a review.