Prefix type radios hidden
The radio buttons used to select the prefix type are now hidden when
allocating from a pool where it is set. The user can still enable them
by clicking the info text displayed instead.

Also added code to handle the case when a pool's default prefix length
is null.

Fixes #102