Fetch less prefixes => improve prefix view
We previously fetched all children of matches to be able to correctly
count the number of childs for assignments. Since we now have logic to
mark the number of children as unknown for certain cases within a type
(ie assignment), we can now remove that join and instead fetch those
prefixes on demand.

This will bring improvements to many searches where previously a lot of
unrelated information would be returned even though we didn't directly
ask for it. In addition, we would sooner hit the result limit and thus
return less relevant prefixes.

Epic win.

Part of #105.