Show prefix type radio buttons
When a prefix is allocated from a pool, the prefix type radio buttons
aer hidden from the user. If the user then instead selected to allocate
a prefix from a prefix or manually, the radio buttons were still hidden.
Made sure that they are shown again when the prefix allocation type is
changed.