Added some auth functionality to the web ui
Made the web ui and NapModel aware of the authentication business. The
foundation is there; users are presented a login screen, the NapModel
has access to user data, but there are still many quirks to iron out.

First and foremost the user data is saved in a class with a shared
state. This probably means that multiple requests share this state and
therefore also shares user data...

Still a good part of #115