Added basic configuration file handling
Created a configuration class which extends the standard ConfigParser
class. The main addition is a shared state; as long as it has been
instanciated and configured once, the next time an instance is created
it will share state with the first one.

napd, which creates the first instance of the config class, also merges
the command line options with the options from the configuration file.

Lastly, made use of the new configuration architecture in napd and the
xmlrpc class.

Part of #119