Add PID file support
This adds support for writing a PID file upon startup and if it already
exists we exit with the assumption of another NIPAPd already running.
This is a fairly safe assumption as we not only check for the files
existance but that it is locked. If nipapd dies or is killed, the lock
will be removed and so no stale pid files can hang around (the file will
actually exist, but won't have the lock..).

The default is not to write a pid file, but one can be specified in the
configuration file or using command line arguments.

Fixes #113