Fix handling of missing configuration file
If the user has specified a configuration file which cannot be found, an
error message is printed to STDERR. We cannot use our regular logger
channel for that, since it hasn't been initialized at this point.

If no config file is specified, we look for our default
(/etc/nipap/nipap.conf) and if we can't find it, we'll just roll with
the built-in defaults.

Fixes #123