Fix config for db + general config fixes
All database parameters are now read from the configuration file instead
of being hardcoded. There is a ugly little workaround for db_host which
needs to not be passed to psycopg2 connect in case one wishes to connect
over a UNIX socket. None would be the best way to handle this in Python,
but since an empty value in the configuration is not intepreted as None,
we do special handling for it and instead interpret '' (ie empty string)
as None for db_host. No other parameters inherit the same handling.
Python and forward support allow_no_value, but since we would like to be
compatible with slightly older versions, we need this hack.

SafeConfigParser is now used instead of ConfigParser since it is
recommended. It is new since 2.3 so we should be all good.

Added a bunch of comments to the configuration file, although it's not
yet complete.. stay tuned :)

Fixes #125