Began working on Java client library
The intent is to provide a library similar to pynipap. I guess the
previous Python implementation shines through; I'm not very
familiar with Java and what best practices apply.

A difference to the pynipap API is that an authentication object is
required for every function call which interacts with the NIPAP server.