Moved files around & added save function to Schema
Moved the files which are in the jnipap package to an own directory and
let the Test-file stay out of it.

Created a save-method in the Schema class; new Schemas can now be
created as well as old ones edited.

Made the NIPAP credentials configurable, they can now be set when the
Connection object is created or set directly on the Connection object.