Debian packaging makeover
This builds two packages out; nipap-common and nipapd. The former holds
some shared libraries while the latter is just the actual daemon
together with its configuration files. The reason for splitting them up
is mainly since nipapd depends on PostgreSQL and on a web front-end,
which require nipap-common, we might not want a database server.

This change was kind of messy so I hope I got it all right.

Part of #120