SQL fiddling
Changed the way the SQL makefile works. It first creates the new nipap
user and then authenticates as it when adding the tables. This way the
tables are automatically owned by the nipap user and we don't have to
add a bunch of GRANT commands. Should ease the installation from debian
configure scripts as well.