Verification of auth backend config
Configuration for each authentication backend is now verified when an
AuthFactory is created. Currently the verification is very static; it
only checks that the options are set in the config file.

The get()-function will throw an exception when a key does not exist.
This exception is not catched, so if an error occurrs the verification
methods will throw an exception defined in the ConfigParser module.