Altered config verification
The config verification is now performed differently; instead of using
special functions, an instance of each auth backend is created. The auth
backends were altered to not perform authentication until explicitly
asked; before it was done in the constructor. This has the side effect
of instance variables not beeing populated until the
authenticate()-method is run on the auth object.