Fixed problem with Inet-type
When running nipapd on Python 2.5 the rows obtained from the ip_net_plan
table could not be directly marshalled as in later versions due to the
display_prefix beeing represented as an Inet-object and not a string. It might
also be related to a previous change when registering the Inet-type;
maybe it was not in use earlier?

Anyway, by explicitly converting the display_prefix to a string the
problem was solved.

Bumped version to 0.5.2.