Moved authentication options into connection obj
Instead of having two objects, one for the authentication options and
one for the connection, the authentication options were added to the
Conenction class. Also, the Connection object was changed from its
current singelton variant to an "ordinary" class. As the authentication
options still were passed to each function which accesses the NIPAP
backend, we could just as well pass the entire connection this way.