Added databas mapped connection class
Created the OConnection class which maps the oracle NIPAP_CONNECTION
object type to the jnipap.Connection class. As we have no control over
how the constructor is called when an object is passed between PL/SQL
and Java, the jnipap.Connection was modified to place the setup logic in
a separate function which is called from the constructor used in the
normal case as well as the readSQL function in the OConnection class.