Altered prefix save methods
Instead of setting the prefix addition options (whether to create from
pool, from prefix and so on) in a hash which was passed to the save
method, a few overloaded save methods have been created, one for each
assignment method. When assigning a prefix fom a pool, a pool object is
simply passed to the save() method. When assigning from a prefix, a
prefix object is passed. When saving changes made to an already existing
prefix or creating a new fully defined one, nothing is passed.

Also, an object holding prefix addition options was created.