Verification of schema argument/attribute
It is now verified that a schema attributes which many of the pynipap
functions take as argument actually contains a Schema object before
trying to access it. The same goes for the save()-functions which rely
on the user having set the schema attribute on the object before calling
save(); now the save method verifies that the schema attribute actually
contains a schema before trying to access it.

Fixes #143