New error page
The NIPAP web UI now has an error page whith a design similar to the
rest of the site. Unfortunately, the actual exception which triggered
the error is not available in the ErrorController, so no relevant
information regarding the error is given...

To make this complete, a more extensive change similar to the error
handling of the XHR controller is needed.

Fixes #136