Add option to specify what address to listen to
--listen makes it possibly for the user to specify one or several
addresses to listen to. The default is now changed to listen to
localhost (127.0.0.1) instead of all interfaces. The port to listen to
can be specified either using --port or per address by appending +
followed by the port, e.g. 127.0.0.1+1234 to listen on 127.0.0.1 and
port 1234.

An example to listen on 127.0.0.1 on standard port 1337 and on 1.3.3.7
on port 6667:
  --listen 127.0.0.1,1.3.3.7+6667

This fixes #132