Fix prefix list search key 'family'
The prefix_list API call search key 'family' did not work as it was
handled as any other prefix property (each one residing in its own
column in the ip_net_plan table). The search query was modified to
calculate the family from the prefix and match against this value
instead.