Improve lookup speed for find_free_prefix
Increase step size which fixes problem where find_free_prefix (ffp)
would spin in a prefix for 2^(wanted_prefix_len - covering_prefix_len)
For example, if we wanted a /112 from a /32 and the first /64 in that
/32 is taken, we would end up spinning through the first /64, stepping
by a /112 at a time. That is 2^112-64 spins, or simply a ridiculously
large number. This improvement instead will step over that entire /64 in
one spin. Much better - hope I didn't break anything :)

Fixes #168