Fix ipv6 prefix search
Fix bug in nipap handling of IPv6 prefixes where bits in the host part
are set. PostgreSQL doesn't like when any bits in the host part is set,
so we need to zero that before it's sent to PostgreSQL. It was done for
IPv4 before and now it's done for IPv6 as well.

Fixes #171