Improvements to popup/prefix menu
The popup menu used to display the edit / remove option for each and
every prefix has been rewritten a bit. It's been renamed to getPopupMenu
which will create a boilerplate menu which can then be appended and
modified for its particular use. This made it possible to put the very
same menu on the schema page.

The overlay used to detect clicks outside of the menu (to trigger
closing of the menu) has been integrated into the getPopupMenu instead
of statically in a template.

All CSS classes and similar have been renamed from prefix_button to
popup_button.

Part of #161