Re-write of insertPrefixList function
The insertPrefixList function is central to the NIPAP web UI as it is
responsible for inserting prefixes into the list of prefixes displayed
to the user. It has also been a quite messy beast, as is needs to handle
many different cases (initial population, population when expanding a
single prefix, inserting data from a "next page" scroll event). Lastly,
it did not work as intended as a couple of input cases were not handled.

The function has now been completely rewritten in a hopefully much more
clear and understandable fashion. Greater care has also been taken to
really cover as many cases as possible.