Improve regexp search for NULL values
We handle empty text strings as NULL in the database most of the time
(should fix that too :) and this results in, for example, a regexp
search for '.*' will not match anything. To resolve this, we now
COALESCE the column with '' for the WHERE clause of our SQL query which
results in a match.

Part of #48