Fix COALESCE search
We do COALESCE for some columns and operators to be able to match a
description with a regexp search for '.*' even when it is NULL. We also
did this for the equal operator (=) which proved to be less successfull,
thus backing out that change.

In a way, part of #191
Also part of #48, where we back out a bit of the change that fixed it.