Move SQL functions into a functions.plsql
calc_indent(), find_free_prefix(), get_prefix() and a bunch of table
trigger functions is moved from their own separate .plsql and
ipnet.plsql file into functions.plsql. Although the separate files have
been very convenient while developing, it is easier to install if there
are less files overall. This reduces it to two.

Also when upgrading from one version of NIPAP to another, if there are
changes to the SQL functions or tables it becomes easier to handle like
this. Functions can be replaced in place, ie just loading the
functions.plsql will overwrite the loaded functions without breaking any
dependancy (ie they are referenced in a trigger on a table). Changes on
to tables on the other hand will need to be handled somewhat different.
The idea is to let the current ip_net.plsql be the base and when we do a
change it will be shipped as an ALTER statement in a separate file so to
go to a new version one would first load the ip_net.plsql and then load
the changes on top of that to get to the latest state.

Fixes #187