Fix the floating prefix list header
This finalises the code on the floating header for the prefix list.
Whenever search results are displayed, we add a header and when the user
scrolls down in the result, the headers "pops loose" and follows the
scorlling at the top of the screen - very nice effect and usefule!

Fixes #170