Remove floating header from Add prefix page
When adding the (floating) header to the prefix list page, it was also
added to the Add prefix page for when the user wants to add a new prefix
from another prefix (from-prefix). In that case the same search page and
result is displayed but as the design of the page is slightly different
the new headers didn't line up very well at all, so I removed them.

The idea is to remove the from-prefix mode on the Add prefix page
anyway, as that case should be handled with a option in the toolbox
menu, see #147.

Fixes #170