Added scrolling function
Added a function which scrolls web page to place element of interest in
view. To avoid annoying the user too much, the function determines
whether the element already is visible, with a 10% margin on top and
bottom of screen, and avoids scrolling if it is.

Otherwise, if the element is placed above the screen, scroll to place it
on top of screen.  Also if the element is larger then viewport, place it
on top to make as much of the item as possible visible.

Else (which should mean that the element is placed below the bottom
of the viewport), scroll to place bottom of item just above bottom of
viewport.

Part of #180