scrollItemVisible function now used
The scrollItemVisible function which implements an "adaptive" scrolling
scheme is now being used on many places in the web ui. Its use has been
deliberately avoided at certain places, when we really want to move the
viewport to focus entirely on one section for a longer period of time,
such as when the user has chosen a prefix allocation method.

Fixes #180