Modify smart_search_prefix for schema-to-vrf
Modify smart_search_prefix and search_prefix to not require the
schema_spec. Searching a VRF is now like any other search parameter. The
idea is to let the user send one part of the query in dictSQL format and
the other as a string to have it interpreted by our 'smart' parsing
functions.

Part of #214