Prefix VRF and pool refs default to natural key
In an effort to decrease the importance of the artificial construct ID
numbers in the NIPAP API, the API has been slightly modified.

Now, an entry in the prefix attribute dict matching the name of the
object type referred to (such as 'vrf' for VRFs and 'pool' for pools)
are interpreted as the natural key of the object instead of the ID. The
natural key for a VRF is the VRF (such as 123:456) and for a pool its
name. The other unique object attributes can be used to refer to a pool
or VRF by using the keys 'vrf_id', 'vrf_name' and 'pool_id' in addition
to the natural 'pool' and 'vrf'.