Renamed VRF property 'vrf' to 'rt' and more...
The VRF property 'vrf' was renamed to 'rt' for clarity and correctness.
In addition to this, all references to other objects in dict keys were
changed to 'object'_'property'; in the output from the 'list_prefix' API
function for example, there are three keys describinf the VRF: 'vrf_id',
'vrf_rt' and 'vrf_name'. This also applies whn sending data to NIPAP,
the same keys can be used to set what VRF a prefix should be in.

In addition to this there has been an effort to hide the artificial
construct of a no-VRF-VRF. For simplicity in maintaining database
consistency, prefixes which are not in a VRF are still 'linked' to an
entry in the VRF-table. This entry is now hidden when searching and
listing VRFs.

Part of #214