Fixed DISTINCT-error and queries with None-values
There was an error in the search_prefix function which manifested itself
in that only one prefix was returned if there were multiple rows in the
ip_net_plan table with the same prefix and different VRFs. As the query
used DISTINCT only on the prefix column, other matching prefixes got
DISTINCT:ed away.

Also added support for None/NULL-values in queries by adding the 'is'
and 'is_not'-operator and automatically changing the 'equals' and
'not_equals'-operators to these if the value of interest is None.