Fix URL updater
The URL update thingy pretty much always used a '?...' URL leading to a
page reload. This was dependant upon if the url.attr('query') was set or
not, but it seems it was in most cases although empty ''. Modified the
if to check so it's not empty in which case it won't add '?' to the URL
and thus not reload the page.