Allow empty search
This allows the user to perform a search for an empty query string,
essentially listing all prefixes. This was previously not possible as
it didn't seem sensible to list all prefixes. Given the VRF selection
and smaller deployments with less prefixes, it might actually make sense
so we allow it.

It is implemented through the "explicit" search option that we have
since before. An explicit search is one initiated by pressing the search
button or pressing enter when focusing the query string input bar. The
automatic searches performed on updates to the query string input are
not explicit.

Fixes #272